slackagent/pyproject.toml
[project]
name = "slackagent"
version = "0.1.0"
description = "Slack bot that forwards emoji-marked threads to a generic A2A agent."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "a2a-sdk>=0.3.0",
    "aiohttp>=3.11.18",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
]
